+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.0.0] - 2024-11-05
 ### Added
+- Time signature in requests, now you can request loops in 3/4, 5/4, 7/8... or whatever is on your mind
 - Macro feature: Now you can trigger a set of commands with different timeouts just using a single command or "macro"
 - aliases.json has a new section called "macros". Not a breaking change because it is fixed automatically by this update
 - New .env flag - SILENCE_MACRO_MESSAGES for new Macro feature. It disables message output for macro commands